server/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
pandas==1.5.3
numpy==1.21.2
APScheduler==3.8.1
plotly==5.3.1
python-dateutil==2.8.2
SQLAlchemy==1.4.23
PyJWT==2.3.0
bcrypt==3.2.0 